refactor(workflows): decompose work.workflow.js into modular step system
Extract work orchestration into 17 reusable step modules (bootstrap, brief,
check, ci, cleanup, commit, complete, follow-up, implement, pr, ready,
reports, spec, task-advance, tasks, ticket, transition) under
workflows/work/steps/, reducing main orchestrator from 1108 to 118 lines.
Create supporting modules (cli.js, plan-generator.js, inspect.js,
transition-step.js, work-helpers.js, tdd-enforcement.js) to enable
dependency injection and testability. Verify generatePlan() output
identical pre/post refactor (104 passing tests, 18 pre-existing failures).
